net: use __int128 for partial value storage in ip checksum
This makes carry handling simpler for the compiler (though it should have
been able to do this without help).

+++ net/ip_checksum.hh
@@ -14,7 +14,7 @@ namespace net {
 uint16_t ip_checksum(const void* data, size_t len);
 
 struct checksummer {
-    uint64_t csum = 0;
+    __int128 csum = 0;
     bool odd = false;
     void sum(const char* data, size_t len);
     void sum(const packet& p);
